feat(version): add version subcommand
