+++ .goreleaser.yml
@@ -9,7 +9,8 @@ builds:
       - darwin
       - windows
       - linux
-
+    ldflags:
+      - -s -w -X main.version={{.Version}}
 archives:
   - replacements:
       darwin: Darwin
